# openapi/LupiraLocationApi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Lupira Location API",
    "description": "Location and presence backend for Lupira. Authenticate with a Bearer token issued by the OIDC provider (Authentik).",
    "version": "v1"
  },
  "paths": {
...
    "/devices": {
      "get": {
        "tags": [
          "Devices"
        ],
        "summary": "List the caller's registered location-tracking devices.",
        "operationId": "ListDevices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "Devices"
        ],
        "summary": "Register a device; returns the one-time ingest API key.",
        "operationId": "RegisterDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDeviceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/devices/{id}": {
      "put": {
        "tags": [
          "Devices"
        ],
        "summary": "Rename a device.",
        "operationId": "RenameDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Devices"
        ],
        "summary": "Retire a device (revokes its ingest keys).",
        "operationId": "RetireDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/ingest/location": {
      "post": {
        "tags": [
          "Ingest"
        ],
        "summary": "Ingest a batch of GPS fixes (NDJSON, one fix per line).",
        "operationId": "IngestLocation",
        "requestBody": {
          "content": {
            "application/x-ndjson": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationIngestReceipt"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/ingest/location/cursor": {
      "get": {
        "tags": [
          "Ingest"
        ],
        "summary": "The device's resume cursor (last accepted seq + ts).",
        "operationId": "GetIngestCursor",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationCursor"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/ingest/location/state": {
      "get": {
        "tags": [
          "Ingest"
        ],
        "summary": "Whether tracking is paused for this device (the uploader should stop collecting if so).",
        "operationId": "GetIngestState",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackingStateDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
    "/location": {
      "delete": {
        "tags": [
          "Location"
        ],
        "summary": "Purge raw fixes + derived docs in a time range (owner erase).",
        "operationId": "PurgeLocationHistory",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "deviceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceDto": {
        "required": [
          "id",
          "kind",
          "label",
          "registeredAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "$ref": "#/components/schemas/DeviceKind"
          },
          "label": {
            "type": "string"
          },
          "externalId": {
            "type": [
              "null",
              "string"
            ]
          },
          "registeredAt": {
            "type": "string",
            "format": "date-time"
          },
          "retiredAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        },
        "description": "A registered device."
      },
      "DeviceKind": {
        "enum": [
          "Phone",
          "Watch",
          "Tracker",
          "Other"
        ],
        "description": "Kind of registered device that feeds location telemetry."
      },
      "IngestReject": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "seq": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "reason": {
            "type": "string"
          }
        },
        "description": "A per-row rejection within an otherwise-accepted batch (permanent — the uploader should drop + log it)."
      },
      "LocationCursor": {
        "required": [
          "deviceId"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "lastSeq": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "lastTs": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        },
        "description": "The resume cursor for a device: the highest accepted seq + its timestamp (from the latest-snapshot table)."
      },
      "LocationIngestReceipt": {
        "required": [
          "submitted",
          "inserted",
          "duplicates",
          "rejected",
          "rejects"
        ],
        "type": "object",
        "properties": {
          "submitted": {
            "type": "integer",
            "format": "int32"
          },
          "inserted": {
            "type": "integer",
            "format": "int32"
          },
          "duplicates": {
            "type": "integer",
            "format": "int32"
          },
          "rejected": {
            "type": "integer",
            "format": "int32"
          },
          "highWaterSeq": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "rejects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IngestReject"
            }
          },
          "paused": {
            "type": "boolean"
          }
        },
        "description": "The outcome of a location ingest batch. Idempotent re-uploads show up as int LocationIngestReceipt.Duplicates; the\n            uploader advances past long? LocationIngestReceipt.HighWaterSeq. When tracking is paused the body is discarded and\n            bool LocationIngestReceipt.Paused is true."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "traceId": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "RFC 9457 problem details."
      },
      "RegisterDeviceRequest": {
        "required": [
          "kind",
          "label"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DeviceKind"
          },
          "label": {
            "type": "string"
          },
          "externalId": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Register a device that will feed location telemetry."
      },
      "RegisterDeviceResponse": {
        "required": [
          "device",
          "keyId",
          "apiKey"
        ],
        "type": "object",
        "properties": {
          "device": {
            "$ref": "#/components/schemas/DeviceDto"
          },
          "keyId": {
            "type": "string",
            "format": "uuid"
          },
          "apiKey": {
            "type": "string"
          }
        },
        "description": "Result of registering a device. string RegisterDeviceResponse.ApiKey is the one-time plaintext ingest credential\n            (`{keyId:N}.{secret}`) — it is shown only here and never retrievable again; only its hash is stored."
      },
      "RenameDeviceRequest": {
        "required": [
          "label"
        ],
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          }
        },
        "description": "Rename a device."
      },
      "TrackingStateDto": {
        "required": [
          "deviceId",
          "paused"
        ],
        "type": "object",
        "properties": {
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "paused": {
            "type": "boolean"
          },
          "pausedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Per-device tracking kill-switch state."
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "OIDC bearer token. Send as `Authorization: Bearer <token>`.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}